Fixed: When migrations loaded in IntegrationTestsBase, the timestamp is being parsed as an Int32 - causing overflow.

+++ MigSharp.NUnit/Integration/IntegrationTestsBase.cs
@@ -36,7 +36,7 @@ namespace MigSharp.NUnit.Integration
             const string pattern = @"Migration(\d+)";
             Migrations.AddRange(typeof(IntegrationTestsBase).Assembly.GetTypes()
                 .Where(t => Regex.IsMatch(t.Name, pattern))
-                .OrderBy(t => int.Parse(Regex.Match(t.Name, pattern).Groups[1].Value, CultureInfo.InvariantCulture))
+                .OrderBy(t => long.Parse(Regex.Match(t.Name, pattern).Groups[1].Value, CultureInfo.InvariantCulture))
                 .Select(t => (IMigration)Activator.CreateInstance(t)));
             Timestamps = new List<long>(Migrations.Select(m => TimestampProvider.GetTimestamp(m.GetType())));
             MigrationOptions.SetGeneralTraceLevel(SourceLevels.All);
